backend/pyproject.toml
[project]
name = "dify-fastapi"
version = "1.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "openai>=2.26.0",
    "pydantic-settings>=2.13.1",
    "python-multipart>=0.0.22",
    "uvicorn",
]
